course_feeling/src/main/java/business/Matiere.java
package business;

import java.util.List;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

import java.util.ArrayList;

public class Matiere {
    
    private String name;
    private List<Sondage> sondages = new ArrayList<Sondage>();
    
    public Matiere() {
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public List<Sondage> getSondages() {
        return sondages;
    }

    public void setSondages(List<Sondage> sondages) {
        this.sondages = sondages;
    }
}
